BitmapOpzet(nietklaar)
