+++ Debug/Core/Src/draw.su
@@ -1,0 +1,3 @@
+../Core/Src/draw.c:10:6:plotLine	48	static
+../Core/Src/draw.c:26:6:plotRect	32	static
+../Core/Src/draw.c:36:6:plotCircle	40	static
