+++ Debug/Core/Src/main.su
@@ -1,7 +1,4 @@
 ../Core/Src/main.c:72:5:main	16	static
 ../Core/Src/main.c:159:6:SystemClock_Config	88	static
 ../Core/Src/main.c:207:1:__io_putchar	16	static
-../Core/Src/main.c:213:6:plotLine	48	static
-../Core/Src/main.c:229:6:plotRect	32	static
-../Core/Src/main.c:239:6:plotCircle	40	static
-../Core/Src/main.c:258:6:Error_Handler	4	static
+../Core/Src/main.c:219:6:Error_Handler	4	static
