+++ Debug/Core/ub_lib/stm32_ub_vga_screen.su
@@ -1,4 +1,3 @@
 ../Core/ub_lib/stm32_ub_vga_screen.c:34:6:UB_VGA_Screen_Init	8	static
 ../Core/ub_lib/stm32_ub_vga_screen.c:66:6:UB_VGA_FillScreen	24	static
 ../Core/ub_lib/stm32_ub_vga_screen.c:82:6:UB_VGA_SetPixel	16	static
-../Core/ub_lib/stm32_ub_vga_screen.c:93:6:UB_VGA_SetPixel3d	16	static
