+++ Debug/objects.list
@@ -1,6 +1,8 @@
 "./Core/Src/bitmap.o"
 "./Core/Src/dma.o"
+"./Core/Src/draw.o"
 "./Core/Src/gpio.o"
+"./Core/Src/imglib.o"
 "./Core/Src/main.o"
 "./Core/Src/stm32f4xx_hal_msp.o"
 "./Core/Src/stm32f4xx_it.o"
